Split configuration for vhost_proxy and dnsmasq services

+++ Vagrantfile
@@ -237,16 +237,8 @@ Vagrant.configure("2") do |config|
     SCRIPT
   end
 
-  # Start system-wide services.
+  # System-wide vhost-proxy service.
   # Containers must define a "VIRTUAL_HOST" environment variable to be recognized and routed by the vhost-proxy.
-  #
-  # Wildcard DNS - Mac configuration:
-  # $ sudo mkdir /etc/resolver
-  # $ echo -e "\n# .drude domain resolution\nnameserver 192.168.10.10" | sudo tee -a  /etc/resolver/drude
-  # Check configuration
-  # $ scutil --dns
-  # $ dig myproject.drude
-  #
   if $vconfig['vhost_proxy']
     config.vm.provision "shell", run: "always", privileged: false do |s|
       s.inline = <<-SCRIPT
@@ -254,7 +246,17 @@ Vagrant.configure("2") do |config|
         docker rm -f vhost-proxy > /dev/null 2>&1 || true
         docker run -d --name vhost-proxy -p 192.168.10.10:80:80 -p 192.168.10.10:443:443 -v /var/run/docker.sock:/tmp/docker.sock \
         blinkreaction/nginx-proxy@sha256:54b17e5298e6f6c1d442c3070f9c53f6250898da097f59786e4bfef8e77863df > /dev/null
-
+      SCRIPT
+    end
+  end
+  
+  # System-wide dnsmasq service.
+  # Wildcard DNS - Mac configuration:
+  # $ sudo mkdir /etc/resolver
+  # $ echo -e "\n# .drude domain resolution\nnameserver 192.168.10.10" | sudo tee -a  /etc/resolver/drude
+  if $vconfig['dnsmasq']
+    config.vm.provision "shell", run: "always", privileged: false do |s|
+      s.inline = <<-SCRIPT
         echo "Starting system-wide DNS service bound to 192.168.10.10:53... "
         docker rm -f dns > /dev/null 2>&1 || true
         docker run -d --name dns -p 192.168.10.10:53:53/udp --cap-add=NET_ADMIN \
